requirements.txt
streamlit==1.32.0
google-genai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.0.0